+++ projects/configs/baselines/test_traj_resnet_lr.py
@@ -514,7 +514,7 @@ optimizer = dict(
     lr=3e-4,
     paramwise_cfg=dict(
         custom_keys={
-            'img_backbone': dict(lr_mult=0.1),
+            'img_backbone': dict(lr_mult=0.5),
         }
     ),
     weight_decay=0.01,