+++ tools/dist_train.sh
@@ -36,4 +36,4 @@ python -m torch.distributed.run \
     $CONFIG \
     --seed 2 \
     --launcher pytorch ${@:3}
-    # --resume work_dirs/EfficientOCF_V1.1_1gpu/epoch_4_lss_only.pth
+    # --resume work_dirs/test_traj/epoch_8_lss_only.pth